requirements.txt
# Core LLM Integration
openai>=1.0.0
litellm[proxy]>=1.77.1

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0

# Memory & Context Management
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.0

# Communication & Notifications
matrix-nio>=0.21.0
slack-sdk>=3.23.0
twilio>=8.10.0
requests>=2.31.0

# Scheduling & Task Management
schedule>=1.2.0
apscheduler>=3.10.0
croniter>=1.4.0
pytz>=2023.3

# Web Interface & APIs
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
jinja2>=3.1.0
python-multipart>=0.0.6

# News & Information Feeds
feedparser>=6.0.10
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
arxiv>=2.1.0
praw>=7.7.0

# Data Processing & Analysis
pandas>=2.1.0
python-dateutil>=2.8.0
regex>=2023.10.0

# Configuration & Environment
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.0

# Logging & Monitoring
loguru>=0.7.0
structlog>=23.2.0
prometheus-client>=0.19.0
psutil>=5.9.0

# Security & Authentication
cryptography>=41.0.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.1.0

# Async & Concurrency
aiohttp>=3.9.0
aioredis>=2.0.0
asyncpg>=0.29.0

# NATS Messaging for Agent Communication
nats-py>=2.0.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Utility Libraries
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
humanize>=4.8.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.23.0
